Project complete
Heavily reduced lines of code by further separating things into more functions. The game itself is in runGame() and is recurred when the event handler is called on the #resetGame button's click. Made one last attempt in HTML to make the game scale better in smaller windows. Could not find a way to make it work, so I am calling this good enough for now.